+++ mahos/meas/tweaker.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 """
-Tweaker for manually-tuned Instrument's ParamDicts.
+Generic tweaker for manually-tunable Instrument's ParamDicts.
 
 .. This file is a part of MAHOS project, which is released under the 3-Clause BSD license.
 .. See included LICENSE file or https://github.com/ToyotaCRDL/mahos/blob/main/LICENSE for details.
@@ -61,7 +61,7 @@ class TweakerClient(StatusClient):
 
 
 class Tweaker(Node):
-    """Tweaker for manually-tuned Instrument's ParamDicts."""
+    """Generic tweaker for manually-tunable Instrument's ParamDicts."""
 
     CLIENT = TweakerClient