+++ tests/fixtures.py
@@ -30,7 +30,7 @@ from mahos.meas.spectroscopy import Spectroscopy, SpectroscopyClient
 from mahos.meas.camera import Camera, CameraClient
 from mahos.meas.state_manager import StateManager, StateManagerClient
 from mahos.meas.tweaker import Tweaker, TweakerClient
-from mahos.meas.chrono import Chrono, ChronoClient
+from mahos.meas.recorder import Recorder, RecorderClient
 
 
 script_dir = path.dirname(path.realpath(__file__))
@@ -40,7 +40,7 @@ gparams_name = "localhost::gparams"
 log_name = "localhost::log"
 manager_name = "localhost::manager"
 tweaker_name = "localhost::tweaker"
-chrono_name = "localhost::chrono"
+recorder_name = "localhost::recorder"
 confocal_name = "localhost::confocal"
 tracker_name = "localhost::tracker"
 odmr_name = "localhost::odmr"
@@ -119,7 +119,7 @@ def gconf():
     local_conf(gconf, camera_name)["poll_timeout_ms"] = 50
     local_conf(gconf, manager_name)["poll_timeout_ms"] = 50
     local_conf(gconf, tweaker_name)["poll_timeout_ms"] = 50
-    local_conf(gconf, chrono_name)["poll_timeout_ms"] = 50
+    local_conf(gconf, recorder_name)["poll_timeout_ms"] = 50
 
     # add conf for dummy
     n = split_name(dummy_name)
@@ -154,8 +154,8 @@ def tweaker_conf(gconf):
 
 
 @pytest.fixture
-def chrono_conf(gconf):
-    return local_conf(gconf, chrono_name)
+def recorder_conf(gconf):
+    return local_conf(gconf, recorder_name)
 
 
 @pytest.fixture
@@ -292,9 +292,9 @@ def tweaker(ctx, gconf):
 
 
 @pytest.fixture
-def chrono(ctx, gconf):
-    proc, shutdown_ev = start_node_proc(ctx, Chrono, gconf, chrono_name)
-    client = ChronoClient(gconf, chrono_name)
+def recorder(ctx, gconf):
+    proc, shutdown_ev = start_node_proc(ctx, Recorder, gconf, recorder_name)
+    client = RecorderClient(gconf, recorder_name)
     yield client
     client.close()
     stop_proc(proc, shutdown_ev)